Remove annotation based replicated entity
