+++ sdk/src/main/java/com/akkaserverless/javasdk/replicatedentity/CommandContext.java
@@ -18,13 +18,8 @@ package com.akkaserverless.javasdk.replicatedentity;
 
 import com.akkaserverless.javasdk.MetadataContext;
 
-/**
- * Context for handling a command.
- *
- * <p>This may be passed to any {@link CommandHandler} annotated element.
- */
-public interface CommandContext
-    extends ReplicatedEntityContext, ReplicatedDataFactory, MetadataContext {
+/** Command context for Replicated Entity. */
+public interface CommandContext extends ReplicatedEntityContext, MetadataContext {
   /**
    * The id of the command. This is an internal ID generated by the proxy, and is unique to a given
    * entity stream. It may be used for debugging purposes.
@@ -41,14 +36,4 @@ public interface CommandContext
    * @return The name of the command.
    */
   String commandName();
-
-  /**
-   * Delete the replicated entity.
-   *
-   * <p>When a replicated entity is deleted, it may not be created again. Additionally, replicated
-   * entity deletion results in tombstones that get accumulated for the life of the cluster. If you
-   * expect to delete replicated entities frequently, it's recommended that you store them in a
-   * single or sharded {@link ReplicatedMap}, rather than individual replicated entities.
-   */
-  void delete();
 }
