+++ tck/src/main/java/com/akkaserverless/javasdk/tck/model/replicatedentity/ReplicatedEntityConfiguredEntity.java
@@ -18,12 +18,12 @@ package com.akkaserverless.javasdk.tck.model.replicatedentity;
 
 import com.akkaserverless.javasdk.replicatedentity.ReplicatedCounter;
 import com.akkaserverless.javasdk.replicatedentity.ReplicatedDataFactory;
-import com.akkaserverless.javasdk.replicatedentity.ReplicatedEntityBase;
+import com.akkaserverless.javasdk.replicatedentity.ReplicatedEntity;
 import com.akkaserverless.javasdk.replicatedentity.ReplicatedEntityContext;
 import com.akkaserverless.tck.model.ReplicatedEntity.Request;
 import com.akkaserverless.tck.model.ReplicatedEntity.Response;
 
-public class ReplicatedEntityConfiguredEntity extends ReplicatedEntityBase<ReplicatedCounter> {
+public class ReplicatedEntityConfiguredEntity extends ReplicatedEntity<ReplicatedCounter> {
 
   public ReplicatedEntityConfiguredEntity(ReplicatedEntityContext context) {}
