+++ tck/src/main/java/com/akkaserverless/javasdk/tck/model/replicatedentity/ReplicatedEntityConfiguredEntityHandler.java
@@ -19,7 +19,7 @@ package com.akkaserverless.javasdk.tck.model.replicatedentity;
 import com.akkaserverless.javasdk.impl.replicatedentity.ReplicatedEntityHandler;
 import com.akkaserverless.javasdk.replicatedentity.CommandContext;
 import com.akkaserverless.javasdk.replicatedentity.ReplicatedCounter;
-import com.akkaserverless.javasdk.replicatedentity.ReplicatedEntityBase;
+import com.akkaserverless.javasdk.replicatedentity.ReplicatedEntity;
 import com.akkaserverless.tck.model.ReplicatedEntity.Request;
 
 public class ReplicatedEntityConfiguredEntityHandler
@@ -30,7 +30,7 @@ public class ReplicatedEntityConfiguredEntityHandler
   }
 
   @Override
-  public ReplicatedEntityBase.Effect<?> handleCommand(
+  public ReplicatedEntity.Effect<?> handleCommand(
       String commandName, ReplicatedCounter data, Object command, CommandContext context) {
     switch (commandName) {
       case "Call":
