+++ tck/src/main/java/com/akkaserverless/javasdk/tck/model/replicatedentity/ReplicatedEntityTckModelEntityHandler.java
@@ -19,7 +19,7 @@ package com.akkaserverless.javasdk.tck.model.replicatedentity;
 import com.akkaserverless.javasdk.impl.replicatedentity.ReplicatedEntityHandler;
 import com.akkaserverless.javasdk.replicatedentity.CommandContext;
 import com.akkaserverless.javasdk.replicatedentity.ReplicatedData;
-import com.akkaserverless.javasdk.replicatedentity.ReplicatedEntityBase;
+import com.akkaserverless.javasdk.replicatedentity.ReplicatedEntity;
 import com.akkaserverless.tck.model.ReplicatedEntity.Request;
 
 public class ReplicatedEntityTckModelEntityHandler
@@ -30,7 +30,7 @@ public class ReplicatedEntityTckModelEntityHandler
   }
 
   @Override
-  public ReplicatedEntityBase.Effect<?> handleCommand(
+  public ReplicatedEntity.Effect<?> handleCommand(
       String commandName, ReplicatedData data, Object command, CommandContext context) {
     switch (commandName) {
       case "Process":
